separated scored set storage
