+++ storage/memory/storage.go
@@ -117,28 +117,6 @@ func (s *storage) GetAllFromSet(key string) ([]string, error) {
 	return result, nil
 }
 
-func (s *storage) GetElementsByScore(key string, score float64, maxElements int) ([]string, error) {
-	s.Log.WithTags(spec.Tags{C: nil, L: "D", O: s, V: 13}, "call GetElementsByScore")
-
-	result, err := s.RedisStorage.GetElementsByScore(key, score, maxElements)
-	if err != nil {
-		return nil, maskAny(err)
-	}
-
-	return result, nil
-}
-
-func (s *storage) GetHighestScoredElements(key string, maxElements int) ([]string, error) {
-	s.Log.WithTags(spec.Tags{C: nil, L: "D", O: s, V: 13}, "call GetHighestScoredElements")
-
-	result, err := s.RedisStorage.GetHighestScoredElements(key, maxElements)
-	if err != nil {
-		return nil, maskAny(err)
-	}
-
-	return result, nil
-}
-
 func (s *storage) GetRandom() (string, error) {
 	s.Log.WithTags(spec.Tags{C: nil, L: "D", O: s, V: 13}, "call GetRandom")
 
@@ -183,28 +161,6 @@ func (s *storage) RemoveFromSet(key string, element string) error {
 	return nil
 }
 
-func (s *storage) RemoveScoredElement(key string, element string) error {
-	s.Log.WithTags(spec.Tags{C: nil, L: "D", O: s, V: 13}, "call RemoveScoredElement")
-
-	err := s.RedisStorage.RemoveScoredElement(key, element)
-	if err != nil {
-		return maskAny(err)
-	}
-
-	return nil
-}
-
-func (s *storage) SetElementByScore(key, element string, score float64) error {
-	s.Log.WithTags(spec.Tags{C: nil, L: "D", O: s, V: 13}, "call SetElementByScore")
-
-	err := s.RedisStorage.SetElementByScore(key, element, score)
-	if err != nil {
-		return maskAny(err)
-	}
-
-	return nil
-}
-
 func (s *storage) SetStringMap(key string, stringMap map[string]string) error {
 	s.Log.WithTags(spec.Tags{C: nil, L: "D", O: s, V: 13}, "call SetStringMap")
 
@@ -222,17 +178,6 @@ func (s *storage) Shutdown() {
 	s.ShutdownOnce.Do(func() {
 		close(s.Closer)
 	})
-}
-
-func (s *storage) WalkScoredSet(key string, closer <-chan struct{}, cb func(element string, score float64) error) error {
-	s.Log.WithTags(spec.Tags{C: nil, L: "D", O: s, V: 13}, "call WalkScoredSet")
-
-	err := s.RedisStorage.WalkScoredSet(key, closer, cb)
-	if err != nil {
-		return maskAny(err)
-	}
-
-	return nil
 }
 
 func (s *storage) WalkSet(key string, closer <-chan struct{}, cb func(element string) error) error {
