+++ storage/redis/scored_set_storage.go
@@ -1,0 +1,173 @@
+package redis
+
+import (
+	"strconv"
+
+	"github.com/cenk/backoff"
+	"github.com/garyburd/redigo/redis"
+
+	"github.com/xh3b4sd/anna/spec"
+)
+
+func (s *storage) GetElementsByScore(key string, score float64, maxElements int) ([]string, error) {
+	s.Log.WithTags(spec.Tags{C: nil, L: "D", O: s, V: 13}, "call GetElementsByScore")
+
+	var result []string
+	var err error
+	action := func() error {
+		conn := s.Pool.Get()
+		defer conn.Close()
+
+		result, err = redis.Strings(conn.Do("ZREVRANGEBYSCORE", s.withPrefix(key), score, score, "LIMIT", 0, maxElements))
+		if err != nil {
+			return maskAny(err)
+		}
+
+		return nil
+	}
+
+	err = backoff.RetryNotify(s.Instrumentation.WrapFunc("GetElementsByScore", action), s.BackOffFactory(), s.retryErrorLogger)
+	if err != nil {
+		return nil, maskAny(err)
+	}
+
+	return result, nil
+}
+
+func (s *storage) GetHighestScoredElements(key string, maxElements int) ([]string, error) {
+	s.Log.WithTags(spec.Tags{C: nil, L: "D", O: s, V: 13}, "call GetHighestScoredElements")
+
+	var result []string
+	var err error
+	action := func() error {
+		conn := s.Pool.Get()
+		defer conn.Close()
+
+		result, err = redis.Strings(conn.Do("ZREVRANGE", s.withPrefix(key), 0, maxElements, "WITHSCORES"))
+		if err != nil {
+			return maskAny(err)
+		}
+
+		return nil
+	}
+
+	err = backoff.RetryNotify(s.Instrumentation.WrapFunc("GetHighestScoredElements", action), s.BackOffFactory(), s.retryErrorLogger)
+	if err != nil {
+		return nil, maskAny(err)
+	}
+
+	return result, nil
+}
+
+func (s *storage) RemoveScoredElement(key string, element string) error {
+	s.Log.WithTags(spec.Tags{C: nil, L: "D", O: s, V: 13}, "call RemoveScoredElement")
+
+	action := func() error {
+		conn := s.Pool.Get()
+		defer conn.Close()
+
+		_, err := redis.Int(conn.Do("ZREM", s.withPrefix(key), element))
+		if err != nil {
+			return maskAny(err)
+		}
+
+		return nil
+	}
+
+	err := backoff.RetryNotify(s.Instrumentation.WrapFunc("RemoveScoredElement", action), s.BackOffFactory(), s.retryErrorLogger)
+	if err != nil {
+		return maskAny(err)
+	}
+
+	return nil
+}
+
+func (s *storage) SetElementByScore(key, element string, score float64) error {
+	s.Log.WithTags(spec.Tags{C: nil, L: "D", O: s, V: 13}, "call SetElementByScore")
+
+	action := func() error {
+		conn := s.Pool.Get()
+		defer conn.Close()
+
+		_, err := redis.Int(conn.Do("ZADD", s.withPrefix(key), score, element))
+		if err != nil {
+			return maskAny(err)
+		}
+
+		return nil
+	}
+
+	err := backoff.RetryNotify(s.Instrumentation.WrapFunc("SetElementByScore", action), s.BackOffFactory(), s.retryErrorLogger)
+	if err != nil {
+		return maskAny(err)
+	}
+
+	return nil
+}
+
+func (s *storage) WalkScoredSet(key string, closer <-chan struct{}, cb func(element string, score float64) error) error {
+	s.Log.WithTags(spec.Tags{C: nil, L: "D", O: s, V: 13}, "call WalkScoredSet")
+
+	action := func() error {
+		conn := s.Pool.Get()
+		defer conn.Close()
+
+		var cursor int64
+
+		// Start to scan the set until the cursor is 0 again. Note that we check for
+		// the closer twice. At first we prevent scans in case the closer was
+		// triggered directly, and second before each callback execution. That way
+		// ending the walk immediately is guaranteed.
+		for {
+			select {
+			case <-closer:
+				return nil
+			default:
+			}
+
+			reply, err := redis.Values(conn.Do("ZSCAN", s.withPrefix(key), cursor, "COUNT", 100))
+			if err != nil {
+				return maskAny(err)
+			}
+
+			cursor, values, err := parseMultiBulkReply(reply)
+			if err != nil {
+				return maskAny(err)
+			}
+
+			for i := range values {
+				select {
+				case <-closer:
+					return nil
+				default:
+				}
+
+				if i%2 != 0 {
+					continue
+				}
+
+				score, err := strconv.ParseFloat(values[i+1], 64)
+				if err != nil {
+					return maskAny(err)
+				}
+				err = cb(values[i], score)
+				if err != nil {
+					return maskAny(err)
+				}
+			}
+
+			if cursor == 0 {
+				break
+			}
+		}
+
+		return nil
+	}
+
+	err := backoff.RetryNotify(s.Instrumentation.WrapFunc("WalkScoredSet", action), s.BackOffFactory(), s.retryErrorLogger)
+	if err != nil {
+		return maskAny(err)
+	}
+
+	return nil
+}
